use correct artifact uris
